requirements.txt
instagrapi==1.18.5
httpx==0.25.1
Pillow>=8.1.1
python-dotenv
openai